+++ examples/react-app/src/routes.jsx
@@ -1,5 +1,6 @@
 import Home from './pages/Home';
 import Accordion from './pages/Accordion';
+import Dimmer from './pages/Dimmer';
 import Avatar from './pages/Avatar';
 import Breadcrumbs from './pages/Breadcrumbs';
 import Affix from './pages/Affix';
@@ -8,6 +9,7 @@ import Autocomplete from './pages/Autocomplete';
 import BackToTop from './pages/BackToTop';
 import Back from './pages/Back';
 import Badge from './pages/Badge';
+import BottomNav from './pages/BottomNav';
 import Button from './pages/Button';
 import Callout from './pages/Callout';
 import Card from './pages/Card';
@@ -34,11 +36,13 @@ import Sidebar from './pages/Sidebar';
 import Skiplinks from './pages/Skiplinks';
 import Sticky from './pages/Sticky';
 import Table from './pages/Table';
+import Timeline from './pages/Timeline';
 import Toggle from './pages/Toggle';
 import Video from './pages/Video';
 import Carousel from './pages/Carousel';
 import Pagination from './pages/Pagination';
 import Notification from './pages/Notification';
+import Tabs from './pages/Tabs';
 import Progress from './pages/Progress';
 
 export const AppRoutes = [
@@ -50,6 +54,7 @@ export const AppRoutes = [
   { path: '/avatar', element: <Avatar />, title: 'Avatar' },
   { path: '/back-to-top', element: <BackToTop />, title: 'Back to Top' },
   { path: '/back', element: <Back />, title: 'Back' },
+  { path: '/bottom-nav', element: <BottomNav />, title: 'Bottom Nav' },
   { path: '/badge', element: <Badge />, title: 'Badge' },
   { path: '/breadcrumbs', element: <Breadcrumbs />, title: 'Breadcrumbs' },
   { path: '/button', element: <Button />, title: 'Button' },
@@ -59,6 +64,7 @@ export const AppRoutes = [
   { path: '/checkbox', element: <Checkbox />, title: 'Checkbox' },
   { path: '/chip', element: <Chip />, title: 'Chip' },
   { path: '/collapse', element: <Collapse />, title: 'Collapse' },
+  { path: '/dimmer', element: <Dimmer />, title: 'Dimmer' },
   { path: '/dropdown', element: <Dropdown />, title: 'Dropdown' },
   { path: '/form', element: <Form />, title: 'Form' },
   { path: '/forward', element: <Forward />, title: 'Forward' },
@@ -82,6 +88,8 @@ export const AppRoutes = [
   { path: '/skiplinks', element: <Skiplinks />, title: 'Skiplinks' },
   { path: '/sticky', element: <Sticky />, title: 'Sticky' },
   { path: '/tables', element: <Table />, title: 'Tables' },
+  { path: '/tabs', element: <Tabs />, title: 'Tabs' },
+  { path: '/timeline', element: <Timeline />, title: 'Timeline' },
   { path: '/toggle', element: <Toggle />, title: 'Toggle' },
   { path: '/video', element: <Video />, title: 'Video' },
 ];