+++ packages/breadcrumbs/stories/it-breadcrumbs.stories.ts
@@ -12,9 +12,24 @@ interface BreadcrumbProps {
 
 const meta = {
   title: 'Componenti/Breadcrumbs',
-  tags: ['alpha', 'web-component', 'a11y-ok'],
+  tags: ['new', 'web-component', 'a11y-ok'],
   component: 'it-breadcrumbs',
-
+  parameters: {
+    a11y: {
+      config: {
+        rules: [
+          {
+            id: 'list',
+            enabled: false,
+          },
+          {
+            id: 'listitem',
+            enabled: false,
+          },
+        ],
+      },
+    },
+  },
   args: {
     'it-aria-label': 'Percorso di navigazione',
     dark: false,