+++ packages/timeline/src/index.ts
@@ -1,0 +1,4 @@
+export { ItTimeline } from './it-timeline.js';
+export { ItTimelineElement } from './it-timeline-element.js';
+export { ItTimelinePoint } from './it-timeline-point.js';
+export * from './types.js';