+++ packages/globals/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @italia/globals
 
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- Aggiornamento della versione
+- Updated dependencies []:
+  - @italia/i18n@1.0.0-alpha.8
+
+## 1.0.0-alpha.7
+
+### Patch Changes
+
+- [#316](https://github.com/italia/dev-kit-italia/pull/316) [`572df4d`](https://github.com/italia/dev-kit-italia/commit/572df4dfe4e249cc98aa6650a2bf8f9726794f40) Thanks [@giuliaghisini](https://github.com/giuliaghisini)! - Added it-toggle component as an extension of it-checkbox
+
+- Updated dependencies []:
+  - @italia/i18n@1.0.0-alpha.7
+
 ## 1.0.0-alpha.6
 
 ### Patch Changes